Added compiler
This is a compiler that uses a modified version of the Pascallite language. Each stage improves upon the last by implementing new features of a compiler system. Stage 0 implements the syntax rules of the language and basic grammar rules of how the structure of the programs are written. Stage 2 adds the ability to create main and assign codes, as well as basic arithmetic. Stage 3 adds comparison functions.