Added files from Artificial Intelligence Class
These AI agents were created to challenge other class member's agents in a series of competitive games.

Off-road-rally is a game where the agent needs to take the best route over from the starting position to the win location, with each location having a set weight. When the best path is found, the search time is taken into account to break ties. The orrGBFS agent uses a greedy best first search to find the outcome. Building from that agent, the orrAgentP utilizes an A* search algorithm to find the best path in a short time.

Chocolate Russian Roulette is a game where each agent tries to make the opponent eat the last piece of chocolate in the bottom left corner. Any piece that the agent eats, it eats any pieces that are to the right or above that piece, creating a square of pieces eaten. The crrAgentP agent uses a min-max algorithm to determine which chocolate square to bite. To improve computation time, a first-best-move function was added to ensure that the best move is utilized without computing the best move with a min-max search.

Matching strategy game is a game where the first agent chooses a tile on a board that has one consonant and one vowel. The opponent then has to choose another tile that has either the same consonant or vowel tile. The aim of the game is to win by having 4 tiles connected either by orthogonal, diagonal, or square patterns. The agent can also win by blocking the opponent from any move. This agent also uses a min-max algorithm to determine the next best move.