Added Theory Of Algorithms Assigments
The Graham Scan code implements code that finds the convex hull of a set of points given to the algorithm. The intersection code utilizes a binary search tree algorithm to determine if the set of lines given to the algorithm contains at least one pair of lines that share an intersection point.